+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.3.0] - 2021/8/10
 
+- Add `solveLinEqHermitian` function
 - Add `ldlDecomp` function
 - Add Windows build support
 - Add cleaning shell script `doClean.sh`